+++ apps/coordinator/src/components/Wallet/TransactionPreview.jsx
@@ -1,6 +1,6 @@
-import React from "react";
+import React, { useMemo } from "react";
 import PropTypes from "prop-types";
-import { connect } from "react-redux";
+import { connect, useSelector } from "react-redux";
 import BigNumber from "bignumber.js";
 import { satoshisToBitcoins } from "@caravan/bitcoin";
 import {
@@ -13,20 +13,156 @@ import {
   TableRow,
   TableCell,
   Grid,
-  Tooltip,
+  Alert,
+  AlertTitle,
+  Chip,
+  Typography,
+  Paper,
 } from "@mui/material";
-import UTXOSet from "../ScriptExplorer/UTXOSet";
+import {
+  CheckCircle as CheckCircleIcon,
+  Warning as WarningIcon,
+  Edit as EditIcon,
+} from "@mui/icons-material";
 import { downloadFile } from "../../utils";
 import UnsignedTransaction from "../UnsignedTransaction";
 import { setChangeOutputMultisig as setChangeOutputMultisigAction } from "../../actions/transactionActions";
-import TransactionAnalysis from "../TransactionAnalysis";
-import ScriptTypeChip from "../ScriptTypeChip";
-import { analyzeTransaction } from "../../hooks/useTransactionAnalysis";
-import WarningAmberIcon from "@mui/icons-material/WarningAmber";
+
+/**
+ * Custom hook to get current signing state
+ * @returns {Object} { signedCount, requiredSigners, isFullySigned, hasPartialSignatures, needsSignatures }
+ */
+export const useSigningState = () => {
+  // Get signature data from Redux store
+  const signatureImporters = useSelector(
+    (state) => state.spend.signatureImporters,
+  );
+  const requiredSigners = useSelector(
+    (state) => state.settings.requiredSigners,
+  );
+
+  /**
+   * Calculate current signature status
+   * Only recalculates when signatureImporters or requiredSigners change
+   */
+  const signingState = useMemo(() => {
+    if (!signatureImporters) {
+      return {
+        signedCount: 0,
+        requiredSigners: requiredSigners || 0,
+        isFullySigned: false,
+        hasPartialSignatures: false,
+        needsSignatures: requiredSigners || 0,
+      };
+    }
+
+    // Count signature importers that have finalized signatures
+    const signedCount = Object.values(signatureImporters).filter(
+      (importer) =>
+        importer.finalized &&
+        importer.signature &&
+        importer.signature.length > 0,
+    ).length;
+
+    const effectiveRequiredSigners = requiredSigners || 0;
+    const isFullySigned = signedCount >= effectiveRequiredSigners;
+    const hasPartialSignatures =
+      signedCount > 0 && signedCount < effectiveRequiredSigners;
+    const needsSignatures = Math.max(0, effectiveRequiredSigners - signedCount);
+
+    return {
+      signedCount,
+      requiredSigners: effectiveRequiredSigners,
+      isFullySigned,
+      hasPartialSignatures,
+      needsSignatures,
+    };
+  }, [signatureImporters, requiredSigners]);
+
+  return signingState;
+};
+
+const SignatureStatus = () => {
+  const {
+    signedCount,
+    requiredSigners,
+    isFullySigned,
+    hasPartialSignatures,
+    needsSignatures,
+  } = useSigningState();
+
+  // We don't render anything if no signatures are present
+  if (signedCount === 0) {
+    return null;
+  }
+
+  return (
+    <Paper elevation={1} sx={{ p: 2, mb: 2 }}>
+      <Box display="flex" alignItems="center" mb={1}>
+        {isFullySigned ? (
+          <CheckCircleIcon color="success" sx={{ mr: 1 }} />
+        ) : (
+          <WarningIcon color="warning" sx={{ mr: 1 }} />
+        )}
+        <Typography variant="h6" component="div">
+          Signature Status
+        </Typography>
+      </Box>
+
+      {/* Status chips */}
+      <Box display="flex" alignItems="center" gap={1} mb={2}>
+        <Chip
+          label={`${signedCount} of ${requiredSigners} signatures`}
+          color={
+            isFullySigned
+              ? "success"
+              : hasPartialSignatures
+                ? "warning"
+                : "default"
+          }
+          variant={isFullySigned ? "filled" : "outlined"}
+        />
+        {isFullySigned && (
+          <Chip label="Fully signed" color="success" size="small" />
+        )}
+        {hasPartialSignatures && (
+          <Chip label="Partially signed" color="warning" size="small" />
+        )}
+      </Box>
+
+      {/* Status-specific alerts */}
+      {hasPartialSignatures && (
+        <Alert severity="info" sx={{ mb: 1 }}>
+          <AlertTitle>Partial signatures detected</AlertTitle>
+          This transaction has {signedCount} out of {requiredSigners} required
+          signatures. You need {needsSignatures} more signature
+          {needsSignatures > 1 ? "s" : ""} to broadcast.
+        </Alert>
+      )}
+
+      {isFullySigned && (
+        <Alert severity="success" sx={{ mb: 1 }}>
+          <AlertTitle>Transaction ready</AlertTitle>
+          This transaction has all {requiredSigners} required signatures and is
+          ready to broadcast.
+        </Alert>
+      )}
+
+      {/* Warning about editing clearing signatures */}
+      {signedCount > 0 && (
+        <Alert severity="warning" icon={<EditIcon />}>
+          <AlertTitle>Important: Editing will clear signatures</AlertTitle>
+          If you edit this transaction (inputs, outputs, or fee), all existing
+          signatures will be cleared and you&apos;ll need to collect signatures
+          again from all signers.
+        </Alert>
+      )}
+    </Paper>
+  );
+};
 
 class TransactionPreview extends React.Component {
   componentDidMount() {
-    // Set up change output multisig details when component loads
     const {
       outputs,
       changeAddress,
@@ -34,7 +170,6 @@ class TransactionPreview extends React.Component {
       changeNode,
       setChangeOutputMultisig,
     } = this.props;
-
     outputs.forEach((output) => {
       if (output.address === changeAddress) {
         setChangeOutputMultisig(changeOutputIndex, changeNode.multisig);
@@ -42,77 +177,128 @@ class TransactionPreview extends React.Component {
     });
   }
 
-  buildOutputRows = () => {
-    const { changeAddress, outputs, inputs, feeRate, addressType, requiredSigners, totalSigners } = this.props;
-    const { walletFingerprinting } = analyzeTransaction({
-      inputs: inputs || [],
-      outputs: outputs || [],
-      feeRate: feeRate || 1,
-      addressType,
-      requiredSigners,
-      totalSigners,
-    });
-    return outputs.map((output, idx) => {
-      const isPoisoned = walletFingerprinting.hasWalletFingerprinting && walletFingerprinting.poisonedOutputIndex === idx;
+  renderAddresses = () => {
+    const addressWithUtxos = this.mapAddresses();
+    return Object.keys(addressWithUtxos).map((address) => {
       return (
-        <TableRow key={output.address} style={isPoisoned ? { background: "#fff3e0" } : {}}>
+        <TableRow key={address}>
           <TableCell>
-            <code>{output.address}</code>
-            {output.address === changeAddress && <small>&nbsp;(change)</small>}
-            {isPoisoned && (
-              <Tooltip title="This output matches your wallet's address type and is likely to be identified as change by an outside observer.">
-                <WarningAmberIcon color="warning" fontSize="small" style={{ marginLeft: 4, verticalAlign: "middle" }} />
-              </Tooltip>
-            )}
+            <code>{address}</code>
           </TableCell>
+          <TableCell>{addressWithUtxos[address].utxos.length}</TableCell>
           <TableCell>
-            <code>{BigNumber(output.amount).toFixed(8)}</code>
-          </TableCell>
-          <TableCell>
-            <ScriptTypeChip scriptType={output.scriptType} />
+            <code>{addressWithUtxos[address].amount.toFixed(8)}</code>
           </TableCell>
         </TableRow>
       );
     });
   };
 
-  buildOutputsTable = () => {
+  renderOutputAddresses = () => {
+    const { changeAddress, outputs } = this.props;
+
+    return outputs.map((output) => {
+      return (
+        <TableRow key={output.address}>
+          <TableCell>
+            <code>{output.address}</code>
+            {output.address === changeAddress ? (
+              <small>&nbsp;(change)</small>
+            ) : (
+              ""
+            )}
+          </TableCell>
+          <TableCell>
+            <code>{BigNumber(output.amount).toFixed(8)}</code>
+          </TableCell>
+        </TableRow>
+      );
+    });
+  };
+
+  renderOutputs = () => {
     return (
       <Table>
         <TableHead>
           <TableRow>
             <TableCell>Address</TableCell>
             <TableCell>Amount (BTC)</TableCell>
-            <TableCell>Script Type</TableCell>
           </TableRow>
         </TableHead>
-        <TableBody>{this.buildOutputRows()}</TableBody>
+        <TableBody>{this.renderOutputAddresses()}</TableBody>
         <TableFooter>
           <TableRow>
             <TableCell>TOTAL:</TableCell>
-            <TableCell>{this.calculateOutputsTotal()}</TableCell>
-            <TableCell />
+            <TableCell>{this.outputsTotal()}</TableCell>
           </TableRow>
         </TableFooter>
       </Table>
     );
   };
 
-  calculateOutputsTotal = () => {
+  renderInputs = () => {
+    const { inputsTotalSats } = this.props;
+
+    return (
+      <Table>
+        <TableHead>
+          <TableRow>
+            <TableCell>Address</TableCell>
+            <TableCell>UTXO count</TableCell>
+            <TableCell>Amount (BTC)</TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>{this.renderAddresses()}</TableBody>
+        <TableFooter>
+          <TableRow>
+            <TableCell colSpan={2}>TOTAL:</TableCell>
+            <TableCell>{satoshisToBitcoins(inputsTotalSats)}</TableCell>
+          </TableRow>
+        </TableFooter>
+      </Table>
+    );
+  };
+
+  mapAddresses = () => {
+    const { inputs } = this.props;
+    return inputs.reduce((mapped, input) => {
+      const mappedAddresses = mapped;
+      const { confirmed, txid, index, amount } = input;
+
+      mappedAddresses[input.multisig.address] = mapped[
+        input.multisig.address
+      ] || {
+        amount: BigNumber(0),
+        utxos: [],
+      };
+      mappedAddresses[input.multisig.address].utxos.push({
+        confirmed,
+        txid,
+        index,
+        amount,
+      });
+      mappedAddresses[input.multisig.address].amount = mapped[
+        input.multisig.address
+      ].amount.plus(BigNumber(input.amount));
+      return mappedAddresses;
+    }, {});
+  };
+
+  outputsTotal = () => {
     const { outputs } = this.props;
     return satoshisToBitcoins(
       outputs.reduce(
-        (sum, output) => sum.plus(BigNumber(output.amountSats || 0)),
+        (total, output) => total.plus(BigNumber(output.amountSats || 0)),
         BigNumber(0),
       ),
     );
   };
 
-  downloadPSBT = (psbtData) => {
-    downloadFile(psbtData, "transaction.psbt");
+  handleDownloadPSBT = (psbtBase64) => {
+    downloadFile(psbtBase64, "transaction.psbt");
   };
 
-  render() {
+  render = () => {
     const {
       feeRate,
       fee,
@@ -120,36 +306,23 @@ class TransactionPreview extends React.Component {
       editTransaction,
       handleSignTransaction,
       unsignedPSBT,
-      inputs,
-      outputs,
     } = this.props;
 
     return (
       <Box>
-        <TransactionAnalysis
-          inputs={inputs || []}
-          outputs={outputs || []}
-          feeRate={feeRate || 1}
-        />
-
         <h2>Transaction Preview</h2>
+
+        {/* Signature Status Section */}
+        <SignatureStatus />
         <UnsignedTransaction />
-
         <h3>Inputs</h3>
-        <UTXOSet
-          inputs={inputs || []}
-          inputsTotalSats={inputsTotalSats}
-          showSelection={false}
-          finalizedOutputs
-        />
-
+        {this.renderInputs()}
         <h3>Outputs</h3>
-        {this.buildOutputsTable()}
-
+        {this.renderOutputs()}
         <Grid container>
           <Grid item xs={4}>
             <h3>Fee</h3>
-            <div>{BigNumber(fee).toFixed(8)} BTC</div>
+            <div>{BigNumber(fee).toFixed(8)} BTC </div>
           </Grid>
           <Grid item xs={4}>
             <h3>Fee Rate</h3>
@@ -160,7 +333,6 @@ class TransactionPreview extends React.Component {
             <div>{satoshisToBitcoins(BigNumber(inputsTotalSats || 0))} BTC</div>
           </Grid>
         </Grid>
-
         <Box mt={2}>
           <Grid container spacing={2}>
             <Grid item>
@@ -188,7 +360,7 @@ class TransactionPreview extends React.Component {
                 <Button
                   variant="contained"
                   color="secondary"
-                  onClick={() => this.downloadPSBT(unsignedPSBT)}
+                  onClick={() => this.handleDownloadPSBT(unsignedPSBT)}
                 >
                   Download Unsigned PSBT
                 </Button>
@@ -198,7 +370,7 @@ class TransactionPreview extends React.Component {
         </Box>
       </Box>
     );
-  }
+  };
 }
 
 TransactionPreview.propTypes = {
@@ -216,18 +388,21 @@ TransactionPreview.propTypes = {
   handleSignTransaction: PropTypes.func.isRequired,
   setChangeOutputMultisig: PropTypes.func.isRequired,
   unsignedPSBT: PropTypes.string.isRequired,
+  signatureImporters: PropTypes.shape({}),
+  requiredSigners: PropTypes.number,
 };
 
-const mapStateToProps = (state) => ({
-  changeOutputIndex: state.spend.transaction.changeOutputIndex,
-  network: state.settings.network,
-  inputs: state.spend.transaction.inputs,
-  outputs: state.spend.transaction.outputs,
-  unsignedPSBT: state.spend.transaction.unsignedPSBT,
-  addressType: state.settings.addressType,
-  requiredSigners: state.settings.requiredSigners,
-  totalSigners: state.settings.totalSigners,
-});
+function mapStateToProps(state) {
+  return {
+    changeOutputIndex: state.spend.transaction.changeOutputIndex,
+    network: state.settings.network,
+    inputs: state.spend.transaction.inputs,
+    outputs: state.spend.transaction.outputs,
+    unsignedPSBT: state.spend.transaction.unsignedPSBT,
+    signatureImporters: state.spend.signatureImporters,
+    requiredSigners: state.settings.requiredSigners,
+  };
+}
 
 const mapDispatchToProps = {
   setChangeOutputMultisig: setChangeOutputMultisigAction,