+++ apps/coordinator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.9.2
+
+### Patch Changes
+
+- [#319](https://github.com/caravan-bitcoin/caravan/pull/319) [`33f8c17`](https://github.com/caravan-bitcoin/caravan/commit/33f8c17400bad7996c6a201cf22582b3d051c27d) Thanks [@ice-009](https://github.com/ice-009)! - Introduced the DustChip for coordinator.
+
+## 1.9.1
+
+### Patch Changes
+
+- [#305](https://github.com/caravan-bitcoin/caravan/pull/305) [`c6f7607`](https://github.com/caravan-bitcoin/caravan/commit/c6f76071e151fd359bc12d05c9e038d32995bf5d) Thanks [@Legend101Zz](https://github.com/Legend101Zz)! - Import Partially Signed PSBT works now
+
 ## 1.9.0
 
 ### Minor Changes